fix: safeConvertToLong crashed on decimal-formatted REST number strings
TastyTrade's REST returns numeric fields as JSON strings, and several
(volume, volume-ext) carry a decimal fraction (e.g. "8370896.451959").
Long.parseLong() throws on that -- caught silently, returning null. This
meant refreshOhlcData's REST-based dayVolume update, and the just-added
volume-ext extraction, both silently no-op'd every single cycle; dayVolume
only ever looked populated because the live DxLink tick path sets it
independently. Parsing as double first and truncating handles both plain
integers and decimals.

+++ src/main/java/com/metradingplat/marketdata/infrastructure/output/external/tastytrade/TastyTradeService.java
@@ -1317,7 +1317,13 @@ public class TastyTradeService {
             return Double.isFinite((double) l) ? l : null;
         }
         try {
-            return Long.parseLong(val.toString());
+            // TastyTrade REST manda campos numericos como strings, y varios
+            // (volume, volume-ext) vienen con fraccion decimal (ej.
+            // "8370896.451959") -- Long.parseLong revienta con eso (confirmado
+            // en vivo: volume-ext nunca se aplicaba, silenciosamente, por este
+            // motivo). Parsear como double y truncar cubre enteros y
+            // decimales por igual.
+            return (long) Double.parseDouble(val.toString());
         } catch (Exception e) {
             return null;
         }
